[common] Move VisitedMask from DiscreteGradient to common folder
