+++ core/base/discreteGradient/DiscreteGradient_Template.h
@@ -17,12 +17,12 @@
 #include <DiscreteGradient.h>
 
 using ttk::SimplexId;
+using ttk::VisitedMask;
 using ttk::dcg::Cell;
 using ttk::dcg::CellExt;
 using ttk::dcg::CriticalPoint;
 using ttk::dcg::DiscreteGradient;
 using ttk::dcg::SaddleSaddleVPathComparator;
-using ttk::dcg::VisitedMask;
 using ttk::dcg::VPath;
 
 template <typename dataType, typename triangulationType>
